# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "calm-server",
    "description": "Wire-format contract between calm-server (Rust) and web-calm (TS). Source of truth for generated TypeScript types.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/threads/{thread_id}/card": {
      "get": {
        "tags": [
          "threads"
        ],
        "summary": "Bridge endpoint: resolve a codex thread_id back to its owning card_id.",
        "description": "Used by neige-codex-bridge to attribute hooks fired by the shared codex\napp-server to the correct card.",
        "operationId": "resolve_card_for_thread",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "description": "Provider thread/session id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "description": "Agent provider: codex or claude; defaults to codex",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owning card for this provider thread/session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadCardResolution"
                }
              }
            }
          },
          "400": {
            "description": "Invalid provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No card is mapped to this thread",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/waves/{wave_id}/codex-cards": {
      "post": {
        "tags": [
          "codex"
        ],
        "operationId": "create_codex_card",
        "parameters": [
          {
            "name": "wave_id",
            "in": "path",
            "description": "Wave id to create the codex card under",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Body required (theme is mandatory; cwd/prompt optional)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCodexCardBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Card + linked terminal created atomically; codex daemon spawned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "404": {
            "description": "Wave not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Body missing required fields (e.g. theme)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Daemon spawn failed (rows are persisted; sweeper reaps within ~60s)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentProvider": {
        "type": "string",
        "enum": [
          "codex",
          "claude"
        ]
      },
      "Card": {
        "type": "object",
        "required": [
          "id",
          "wave_id",
          "kind",
          "sort",
          "payload",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "deletable": {
            "type": "boolean",
            "description": "Issue #229 PR A — system-card guard. `true` for user-facing cards\n(the default; all pre-#229 rows backfill via the column DEFAULT in\nmigration 0013). `false` for kernel-owned cards that the user\ncannot remove via REST / plugin callbacks — currently spec cards\n(retroactively undeletable via the same migration's UPDATE) and\nPR B's wave-report cards.\n\n`#[serde(default = \"default_deletable\")]` so wire payloads emitted\nbefore #229 landed (event-log replay fixtures, old test seeds)\nparse as `true` without forcing a fixture rewrite — matches the\nDB DEFAULT (1) in migration 0013. The default-fn lives below\nbecause `bool::default()` would give `false` (the *un*safe\nfallback for a deny-by-omission auth bit)."
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "description": "`\"terminal\"` for built-in PTY cards, `\"ui://<plugin>/<view>\"` for\nplugin-provided cards (the canonical MCP Apps resource URI). The\nkernel never interprets beyond that prefix. `[legacy]`\n`\"plugin:<plugin-id>:<view-id>\"` may also appear on rows persisted\nbefore the M4 cut-over and in server-side perms/manifest enforcement\n— see `docs/architecture/terminology-glossary.md` (plugin card kind)."
          },
          "payload": {
            "type": "object",
            "description": "Opaque JSON blob — ts-rs would otherwise emit `unknown` via the\n`serde-json-impl` feature, but we pin it explicitly so a future\nfeature-flag change can't silently widen / narrow the surface."
          },
          "runtime": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CardRuntimeView"
              }
            ]
          },
          "sort": {
            "type": "number",
            "format": "double"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          },
          "wave_id": {
            "type": "string"
          }
        }
      },
...
      "CardRole": {
        "type": "string",
        "description": "Wave-as-Actor PR3 (#136): authorization label persisted on each card.\n\nThe role decides whether the card's implicit actor (the AI agent bound\nto it, or the user when no agent is bound) is allowed to emit a given\nevent. The gate is checked at the single write entry — see\n`role_gate::enforce_role` — *inside* the transaction, before the event\nrow is appended. Violations roll the txn back; nothing is broadcast.\n\n  * [`CardRole::Spec`] (PR6) is the wave's spec card. Only spec cards\n    may emit `WaveUpdated`; this is the structural choke point that\n    keeps AI workers from rewriting wave-level metadata.\n  * [`CardRole::Worker`] is the default for user-facing card inserts\n    and dispatcher-spawned worker cards. Its events are scoped to the\n    card itself and never broaden.\n  * [`CardRole::ReportCard`] (#229 PR A) is the wave's auto-generated\n    report card. Same kernel-ownership profile as `Spec` — minted by\n    the wave-create path (PR B), one per wave (partial unique index\n    in migration 0013), undeletable from REST / plugin-callback paths.\n    Role-gate-wise it behaves like `Worker`: it only emits `CardUpdated`\n    for its own scope; it does **not** emit `WaveUpdated` (only `Spec`\n    does — preserving the #136 contract).\n\nPersisted as a lowercase string in `cards.role` (migration 0008). The\nserde + sqlx `rename_all = \"lowercase\"` keeps the wire / storage shape\nstable; ts-rs exports the matching TS union (`\"spec\" | \"worker\" |\n\"reportcard\"`) into `web/src/api/generated-events.ts` so the\nfrontend can adopt the enum once any UI lands.",
        "enum": [
          "worker",
          "spec",
          "reportcard"
        ]
      },
      "CardRuntimeView": {
        "type": "object",
        "description": "Live runtime projection joined from the `runtimes` table when a card is\nfetched or serialized.\n\nThis view is not part of the idempotency contract: across retries the\nruntime row may have advanced, so `Card.runtime` may differ between the\nfirst POST response and a retry POST response returning the same operation\nresult. Future cleanup (#581 item 4) will remove the legacy payload-key\nprojection; this typed view is the forward-compatible reader path.",
        "required": [
          "runtime_id",
          "kind",
          "status"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/WorkerSessionKind"
          },
          "provider": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentProvider"
              }
            ]
          },
          "runtime_id": {
            "type": "string"
          },
          "session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/WorkerSessionState"
          },
          "terminal_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "thread_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "thread_status": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "JSON shape returned for every error response — `{error, code}`.\nMirrors the body produced by `CalmError::into_response`. Hand-written\nduplicate of the in-line `json!` body so OpenAPI consumers see a\nconcrete schema.",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code — one of `not_found`, `conflict`,\n`idempotency_collision`, `bad_request`, `unauthorized`,\n`forbidden`, `plugin_install`, `plugin_permission`,\n`plugin_conflict`, `plugin_kernel_too_old`,\n`spec_harness_dormant`, `db_error`, `io_error`, `serde_error`,\n`codex_app_server`, `service_unavailable`, `internal`,\n`forbidden_tool`, `not_a_card_tool`, `tool_call_failed`."
          },
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "NewCodexCardBody": {
        "type": "object",
        "description": "Body for `POST /api/waves/:wave_id/codex-cards`.\n\nDeliberately omits `kind` (always `\"codex\"`) and `payload` (the kernel\npersists schema/UI fields and projects identity from `runtimes`). Empty\n`cwd` falls back to `$HOME` then the server's cwd.\n\n`prompt` is the hands-free entry point: when non-empty, the kernel starts\na shared thread, binds it to the runtime row, sends the prompt via\n`turn/start`, waits for `turn/started` or `turn/completed`, and starts\nthe TUI as `codex resume <thread_id> --remote unix://...`.\n\nEmpty / absent `prompt` reverts to the user-initiated flow: codex\nboots through the shared remote, the composer is empty, the user types\nand hits Enter.\n\nNote: the old `initial_prompt` field (which had been a documented\nno-op since the codex-TUI port) was removed; serde rejects unknown\nfields with the default config, so a stale caller that still sends\nit will get a 422 — that's the intended fail-loud signal to update\nthe caller. The interactive `prompt` channel is the one place\ncallers should be putting text now.\n\n`theme` is required end-to-end (#177): callers MUST send the host\nbrowser's current foreground/background RGB. The renderer uses it so\ncodex's OSC 10/11 startup probe gets matching colors. Forcing it at\nthe type layer means a\ncaller that forgets — the exact bug that motivated this refactor —\nfails at compile time (TS) or at the deserialize step (Rust/JSON,\n422). No `Option`, no `#[serde(default)]`, no implicit fallback.",
        "required": [
          "theme"
        ],
        "properties": {
          "cwd": {
            "type": [
              "string",
              "null"
            ],
            "description": "Working directory codex runs in. Empty string or missing → `$HOME`\n(then `cwd` of server)."
          },
          "icon_bg": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional card-head logo background CSS color. Empty string is ignored."
          },
          "icon_fg": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional card-head logo foreground CSS color. Empty string is ignored."
          },
          "prompt": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hands-free seed prompt. When set and non-empty, codex boots with\nits composer pre-filled and the kernel auto-submits the composer\nonce codex's session is constructed. See the struct doc for the\nfull mechanism."
          },
          "sort": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Sort order within the wave. `None` defaults to \"append to end\"."
          },
          "theme": {
            "$ref": "#/components/schemas/RequestTheme",
            "description": "Host browser's current theme RGB (#177). Required so the terminal\nmodel answers codex's OSC 10/11 startup probe with colors matching\nthe host theme. A caller that omits this field gets 422."
          }
        },
        "additionalProperties": false
      },
...
      "RequestTheme": {
        "type": "object",
        "description": "Wire shape of `NewCodexCardBody.theme` / `NewWave.theme`. Matches the\n`calm_session::TerminalTheme` value type one-for-one — duplicated\nhere so the route can keep its own `ToSchema` derive (the\n`calm_session` crate is utoipa-free).",
        "required": [
          "fg",
          "bg"
        ],
        "properties": {
          "bg": {
            "type": "array",
            "items": false,
            "prefixItems": [
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              }
            ]
          },
          "fg": {
            "type": "array",
            "items": false,
            "prefixItems": [
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "ThreadCardResolution": {
        "type": "object",
        "required": [
          "thread_id",
          "card_id",
          "role"
        ],
        "properties": {
          "card_id": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/CardRole"
          },
          "thread_id": {
            "type": "string"
          },
          "wave_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "WorkerSessionKind": {
        "type": "string",
        "enum": [
          "terminal",
          "codex",
          "claude",
          "shared-spec"
        ]
      },
      "WorkerSessionState": {
        "type": "string",
        "description": "Session state machine column (`worker_sessions.state`, issue #679 §1).\nSingle runtime/session state vocabulary (`worker_sessions.state`, issue #679 §1).",
        "enum": [
          "starting",
          "running",
          "idle",
          "turn_pending",
          "exited",
          "failed",
          "superseded"
        ]
      }
    }
  },
...
}